+++ main.rb
@@ -1,48 +1,9 @@
 require_relative 'app'
+require_relative 'registerpeople'
 
-class Main
+class RentBook
   def initialize
     @app = App.new
-  end
-
-  def options
-    puts 'Please choose an option by entering a number'
-    puts '1 - List all books'
-    puts '2 - List all people'
-    puts '3 - Create a person'
-    puts '4 - Create a book'
-    puts '5 - Create a rental'
-    puts '6 - List all rentals for a given person id'
-    puts '7 - Exit'
-    print 'Option: '
-  end
-
-  def register_people
-    print 'Do you want to create a student (1) or a Teacher (2)? [Input the number]:'
-    selection = gets.chomp
-    num = selection.to_i
-    case num
-    when 1
-      print 'Age: '
-      age = gets.chomp.to_i
-      print 'Name: '
-      name = gets.chomp
-      print 'Has parent permission? [Y/N]: '
-      parent_permission = gets.chomp.downcase == 'y'
-      print 'Classroom: '
-      classroom = gets.chomp
-      @app.create_student(classroom, age, name, parent_permission)
-    when 2
-      print 'Age: '
-      age = gets.chomp.to_i
-      print 'Name: '
-      name = gets.chomp
-      print 'Specialization: '
-      specialization = gets.chomp
-      @app.create_teacher(age, name, specialization)
-    else
-      puts 'Wrong Input.'
-    end
   end
 
   def rent_book
@@ -59,6 +20,24 @@ class Main
     print 'Date: '
     date = gets.chomp
     @app.rent_a_book(date, person_number, book_number)
+  end
+end
+
+class Main
+  def initialize
+    @app = App.new
+  end
+
+  def options
+    puts 'Please choose an option by entering a number'
+    puts '1 - List all books'
+    puts '2 - List all people'
+    puts '3 - Create a person'
+    puts '4 - Create a book'
+    puts '5 - Create a rental'
+    puts '6 - List all rentals for a given person id'
+    puts '7 - Exit'
+    print 'Option: '
   end
 
   def register_book
@@ -81,11 +60,11 @@ class Main
       when 2
         @app.list_people
       when 3
-        register_people
+        Register.new.register_people
       when 4
         register_book
       when 5
-        rent_book
+        RentBook.new.rent_book
       when 6
         print 'ID of person: '
         id = gets.chomp.to_i
